Move Action, PolicyNetOutput, TrainingParameters to common
