+++ rl_2048/bin/playRL2048_jax_dqn.py
@@ -17,12 +17,13 @@ from flax.training.checkpoints import PyTree, restore_checkpoint
 from jax import Array
 from jax import random as jrandom
 
+from rl_2048.dqn.common import Action
 from rl_2048.dqn.jax.dqn import (
     DQN,
     TrainingParameters,
 )
 from rl_2048.dqn.jax.net import PREDEFINED_NETWORKS, load_predefined_net
-from rl_2048.dqn.jax.replay_memory import Action, Transition
+from rl_2048.dqn.jax.replay_memory import Transition
 from rl_2048.dqn.jax.utils import flat_one_hot
 from rl_2048.game_engine import GameEngine, MoveResult
 from rl_2048.tile import Tile
